Fix mode issue
