+++ src/DocumentFormat.OpenXml.Framework/Features/PackageFeature.cs
