+++ src/DocumentFormat.OpenXml.Framework/Packaging/PackageBase.cs
@@ -6,16 +6,17 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Packaging;
+using System.Linq;
 
 namespace DocumentFormat.OpenXml.Packaging;
 
 internal abstract class PackageBase : IPackage
 {
-    private readonly Dictionary<Uri, PackagePartImpl> _parts = new();
+    private readonly Dictionary<Uri, PackagePart> _parts = new();
 
     protected abstract Package Package { get; }
 
-    private IPackagePart GetOrCreatePart(PackagePart part)
+    private IPackagePart GetOrCreatePart(System.IO.Packaging.PackagePart part)
     {
         if (_parts.TryGetValue(part.Uri, out var existing))
         {
@@ -23,7 +24,7 @@ internal abstract class PackageBase : IPackage
             return existing;
         }
 
-        var newItem = new PackagePartImpl(this, part);
+        var newItem = new PackagePart(this, part);
         _parts[part.Uri] = newItem;
         return newItem;
     }
@@ -71,5 +72,37 @@ internal abstract class PackageBase : IPackage
     {
         _parts.Remove(uri);
         Package.DeletePart(uri);
+    }
+
+    private sealed class PackagePart : IPackagePart
+    {
+        private readonly PackageBase _package;
+
+        public PackagePart(PackageBase package, System.IO.Packaging.PackagePart part)
+        {
+            _package = package;
+            Part = part;
+        }
+
+        public System.IO.Packaging.PackagePart Part { get; }
+
+        public IPackage Package => _package;
+
+        public Uri Uri => Part.Uri;
+
+        public string ContentType => Part.ContentType;
+
+        public void DeleteRelationship(string id) => Part.DeleteRelationship(id);
+
+        public IEnumerable<IPackageRelationship> GetRelationships() => Part.GetRelationships().Select(_package.GetOrCreate);
+
+        public Stream GetStream(FileMode open, FileAccess write) => Part.GetStream(open, write);
+
+        public IPackageRelationship CreateRelationship(Uri targetUri, TargetMode targetMode, string relationshipType, string? id)
+            => _package.GetOrCreate(Part.CreateRelationship(targetUri, targetMode, relationshipType, id));
+
+        public bool RelationshipExists(string relationship) => Part.RelationshipExists(relationship);
+
+        public IPackageRelationship GetRelationship(string id) => _package.GetOrCreate(Part.GetRelationship(id));
     }
 }
